Completed simple updates Jeremy suggested
