+++ src/wrapper.js
@@ -1,6 +1,6 @@
 import './util/object-assign';
 import merge from './util/merge';
-import defaultSettings, { syncSettings, syncControlInputs, controlInputs } from './defaultSettings';
+import defaultSettings, { syncSettings, syncControlInputs } from './defaultSettings';
 import { createControls, createTable } from 'webcharts';
 import onInit from './onInit';
 import onLayout from './onLayout';
@@ -9,7 +9,7 @@ import onDraw from './onDraw';
 export default function raveXplorer(element, settings) {
     const mergedSettings = merge(defaultSettings, settings), //Merge user settings onto default settings.
         syncedSettings = syncSettings(mergedSettings), //Sync properties within merged settings, e.g. data mappings.
-        syncedControlInputs = syncControlInputs(controlInputs, syncedSettings), //Sync merged settings with controls.
+        syncedControlInputs = syncControlInputs(syncedSettings), //Sync merged settings with controls.
         controls = createControls(element, {
             location: 'top',
             inputs: syncedControlInputs
